requirements.txt
boto3~=1.36.26
pandas~=2.2.3
flask~=3.1.0
faiss-cpu
sentence-transformers
